combine Slider component into VideoController file
