--- packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/VideoController/components/Slider/index.js
@@ -1,1 +1,0 @@
-export { default } from './Slider'
